fix type errors due to renaming and moving types and files
