+++ dotcom-rendering/src/components/FootballCompetitionSelect.tsx
@@ -1,6 +1,7 @@
 import { Option, Select } from '@guardian/source/react-components';
-import type { FootballMatchKind, Region } from '../footballMatches';
+import type { FootballMatchKind } from '../footballMatches';
 import { palette } from '../palette';
+import { Region } from '../footballDataPage';
 
 type FootballSelectKind = FootballMatchKind | 'Tables';
