+++ dotcom-rendering/src/footballDataPage.ts
@@ -1,9 +1,9 @@
-import { FEFootballPageConfig } from './feFootballDataPage';
-import { FootballMatches, FootballMatchKind } from './footballMatches';
+import type { FootballMatches, FootballMatchKind } from './footballMatches';
 import { FootballTables } from './footballTables';
-import { EditionId } from './lib/edition';
-import { NavType } from './model/extract-nav';
-import { FooterType } from './types/footer';
+import type { FEFootballPageConfig } from './frontend/feFootballDataPage';
+import type { EditionId } from './lib/edition';
+import type { NavType } from './model/extract-nav';
+import type { FooterType } from './types/footer';
 
 export type Region = {
 	name: string;
