+++ dotcom-rendering/src/lib/contributions.ts
@@ -8,7 +8,7 @@ import type { HeaderPayload } from '@guardian/support-dotcom-components/dist/dot
 import { useEffect, useState } from 'react';
 import { hideSupportMessaging } from '../client/userFeatures/cookies/hideSupportMessaging';
 import { userBenefitsDataIsUpToDate } from '../client/userFeatures/cookies/userBenefitsExpiry';
-import type { FootballMatchListPage } from '../footballMatches';
+import type { FootballMatchListPage } from '../footballDataPage';
 import type { ArticleDeprecated } from '../types/article';
 import type { Front } from '../types/front';
 import type { DCRNewslettersPageType } from '../types/newslettersPage';
