+++ dotcom-rendering/src/server/render.footballDataPage.web.tsx
@@ -1,6 +1,7 @@
 import { isString } from '@guardian/libs';
 import { ConfigProvider } from '../components/ConfigContext';
 import { FootballDataPage } from '../components/FootballDataPage';
+import type { FootballMatchListPage, Region } from '../footballDataPage';
 import type { FootballMatchKind } from '../footballMatches';
 import {
 	ASSET_ORIGIN,
@@ -13,7 +14,6 @@ import { polyfillIO } from '../lib/polyfill.io';
 import { createGuardian } from '../model/guardian';
 import type { Config } from '../types/configContext';
 import { htmlPageTemplate } from './htmlPageTemplate';
-import { FootballMatchListPage, Region } from '../footballDataPage';
 
 const fromTheGuardian =
 	'from the Guardian, the world&#x27;s leading liberal voice';
